Allow fractional food quantities on diet ingestas (#602)
* Allow fractional food quantities on diet ingestas.

Nutritionists can enter cooking fractions instead of +/- portions, and
macros update live without a page reload.


* Avoid PMD NullAssignment when parsing invalid food quantities.


---------

+++ src/main/java/com/nutriconsultas/ai/AiDraftIngestaMaterializer.java
@@ -139,7 +139,7 @@ public class AiDraftIngestaMaterializer {
 		final AlimentoIngesta alimentoIngesta = new AlimentoIngesta();
 		alimentoIngesta.setName(alimento.getNombreAlimento());
 		alimentoIngesta.setAlimento(alimento);
-		alimentoIngesta.setPortions(portions);
+		alimentoIngesta.setPortions((double) portions);
 		alimentoIngesta.setUnidad(alimento.getUnidad());
 		if (alimento.getEnergia() != null) {
 			alimentoIngesta.setEnergia(alimento.getEnergia() * portions);
