+++ src/main/java/com/nutriconsultas/mobile/dto/DietAlimentoDto.java
@@ -7,7 +7,7 @@ import com.nutriconsultas.dieta.AlimentoIngesta;
  * Patient-facing food item within a meal slot for mobile diet plan detail (#94, #354).
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record DietAlimentoDto(String nombre, Integer porciones, Integer kcal, String unidad, Double proteina,
+public record DietAlimentoDto(String nombre, Double porciones, Integer kcal, String unidad, Double proteina,
 		Double carbohidratos, Double grasas) {
 
 	public static DietAlimentoDto fromEntity(final AlimentoIngesta alimento) {
